+++ README.md
@@ -151,7 +151,7 @@
     - [x] MenuProduct
     - [x] Order
     - [x] OrderLineItem
-    - [ ] OrderTable
+    - [x] OrderTable
     - [ ] TableGroup
