+++ src/main/java/kitchenpos/ordertable/repository/OrderTableRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.repository;
+package kitchenpos.ordertable.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.ordertable.model.OrderTable;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
     List<OrderTable> findAllByTableGroupId(Long tableGroupId);
