+++ src/test/java/kitchenpos/order/model/OrderVerifierTest.java
@@ -10,11 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.order.model.Order;
+import kitchenpos.ordertable.model.OrderTable;
 import kitchenpos.orderline.model.OrderLineItem;
-import kitchenpos.order.model.OrderStatus;
-import kitchenpos.order.model.OrderVerifier;
 
 class OrderVerifierTest {
     @DisplayName("결제 완료되지 않은 테이블이 존재하면 예외 발생")
